requirements.txt
scikit-learn
pandas
numpy
matplotlib
plotly
mlflow
jupyter
notebook
python-box
pyYAML
ensure
joblib
fastapi
gdown
uvicorn
pydantic
python-multipart
xgboost
python-dotenv
streamlit
zenml[mlflow]
zenml[server]
black
isort
black[jupyter]
-e .
